+++ clientBroker/lib/src/functions.cpp
@@ -1,116 +1,6 @@
+#include "functions.hpp"
 #include "param.hpp"
-
-#include <Arduino.h>
-#include <stdlib.h>
-#include <stdint.h>
-#include <WiFi.h>
-#include <AsyncTCP.h>
-#include <ESPAsyncWebServer.h>
-#include <PubSubClient.h>
-#include <Wire.h>
-#include <DHT.h>
-
 #include <AsyncElegantOTA.h>
-
-extern DHT dht;
-extern AsyncWebServer server;
-extern WiFiClient espClient;
-extern PubSubClient client;
-
-extern const char *mqtt_server;
-
-extern long lastMsg;
-extern char msg[50];
-extern int value;
-
-extern float nivelLuz;
-extern float humedad;
-extern float humedadSuelo;
-extern float temperature;
-extern int tiempoMuestras;
-extern int pesoMuestras;
-//---------------------------------
-
-extern uint8_t tempArray[20];
-extern uint8_t N_fil;
-extern uint8_t current_temp; // Temperatura actual s
-extern uint8_t prom;         // Promedio
-extern uint8_t humeArray[20];
-extern uint8_t current_hume; // Temperatura actual s
-extern uint8_t promhume;     // Promedio
-
-void start_ota_webserver(void);
-void callback(char *topic, byte *message, unsigned int length);
-void reconnect();
-void changeState(String messageTemp, int pin);
-double prome(uint8_t temp[], uint8_t N_filter);
-void pushData(uint8_t *tempArray, uint8_t newTemp, uint8_t N_fil);
-void mandarDatos(const int Read, uint8_t *datoArray, uint8_t N_fil, const char *topic, int min, int max);
-
-DHT dht(DHTPIN, DHTTYPE);
-
-void setup(void) {
-  pinMode(LED_ONBOARD, OUTPUT);
-  Serial.begin(115200);
-
-  start_ota_webserver();  
-  client.setServer(mqtt_server, 1883);
-  client.setCallback(callback);
-
-  pinMode(LED_1, OUTPUT);
-  pinMode(LED_2, OUTPUT);
-  pinMode(LED_3, OUTPUT);
-
-  ledcSetup(PWM_LED_CHANNEL, PWM_FREQ, PWM_RESOLUTION);
-  ledcAttachPin(PWM_LED, PWM_LED_CHANNEL);
-  dht.begin();
-}
-
-void loop()
-{
-  if (!client.connected())
-  {
-    reconnect();
-  }
-  client.loop();
-
-  long now = millis();
-  if (now - lastMsg > tiempoMuestras * DELAY * pesoMuestras) // 1000ms de muestreo
-  {
-    lastMsg = now;
-    char humString[8];
-    char tempString[8];
-    mandarDatos(ANALOG_1, tempArray, N_fil, "esp32/nivelLuz", 0, 4095);
-    mandarDatos(ANALOG_2, humeArray, N_fil, "esp32/humedadSuelo", 880, 1540); //2370, 4095);
-
-    humedad = dht.readHumidity();
-    dtostrf(humedad, 1, 2, humString);
-    client.publish("esp32/humidity", humString); // esp32/humidity
-
-    temperature = dht.readTemperature();
-    dtostrf(temperature, 1, 2, tempString);
-    client.publish("esp32/temperature", tempString); // esp32/temperature
-
-    digitalWrite(LED_ONBOARD, !digitalRead(LED_ONBOARD));
-    // Test fede
-    //Serial.println("");
-    //Serial.print("Sensor Humedad de Suelo: ");
-    //Serial.println(analogRead(ANALOG_2));
-  }
-}
-
-/*
-void loop(void) {
-    digitalWrite(LED, !digitalRead(LED));
-    //humedad = analogRead(HUMIDITY_SENSOR);
-    //Serial.println("");
-    //Serial.print("Sensor de humedad: ");
-    //Serial.println(humedad);
-    delay(DELAY);
-}
-*/
-
-//---------------------------------------FUNCIONES---------------------------------
 
 AsyncWebServer server(80);
 WiFiClient espClient;
@@ -326,4 +216,3 @@ void mandarDatos(const int Read, uint8_t *datoArray, uint8_t N_fil, const char *
   dtostrf(prom, 1, 2, datoString);
   client.publish(topic, datoString);
 }
-
